added seeder and custom page

+++ urban-cart-backend/src/main/java/com/example/urbancart/service/ProductService.java
@@ -1,5 +1,6 @@
 package com.example.urbancart.service;
 
+import com.example.urbancart.common.CustomPage;
 import com.example.urbancart.model.Product;
 import com.example.urbancart.repository.ProductRepository;
 import java.util.UUID;
@@ -8,7 +9,6 @@ import org.springframework.cache.annotation.CacheConfig;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
@@ -27,11 +27,15 @@ public class ProductService {
     this.productRepository = productRepository;
   }
 
-  public Page<Product> findAll(int page, int size, String sortBy, String sortDirection) {
+  public CustomPage<Product> findAll(
+      int page, int size, String sortBy, String sortDirection, Boolean isDeleted, String search) {
     Sort.Direction direction =
         sortDirection.equalsIgnoreCase("asc") ? Sort.Direction.ASC : Sort.Direction.DESC;
     Pageable pageable = PageRequest.of(page, size, Sort.by(direction, sortBy));
-    return productRepository.findAll(pageable);
+    var data =
+        productRepository.findAllByIsDeletedAndNameContainingIgnoreCase(
+            pageable, isDeleted, search);
+    return new CustomPage<Product>(data);
   }
 
   @Cacheable(key = "#id", unless = "#result == null", value = "product")
@@ -56,7 +60,8 @@ public class ProductService {
   }
 
   @CacheEvict(key = "#id", value = "product")
-  public void remove(UUID id) {
-    this.productRepository.deleteById(id);
+  public void remove(UUID id, Boolean isHardDelete) {
+    if (isHardDelete) this.productRepository.deleteById(id);
+    else this.productRepository.softDeleteById(id);
   }
 }
